aiida-cfdemcoupling/pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-cfdem"
dynamic = ["version"]  # read from aiida_aspherix/__init__.py
description = "AiiDA demo plugin that wraps the `aspherix` executable "
authors = [{name = "The AiiDA Team"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
    "Framework :: AiiDA"
]
keywords = ["aiida", "plugin"]
requires-python = ">=3.8"
dependencies = [
    "aiida-core>=2.0,<3"
]


[project.entry-points."aiida.calculations"]
"cfdem.run" = "aiida_cfdem.calculations.cfdem.run:CfdemRun"


[project.entry-points."aiida.parsers"]
"cfdem" = "aiida_cfdem.parsers:AspherixParser"
